oss-deploy-a-fleet-intelligence-solution-for-taxis/scripts/requirements.txt
# SF Taxi Fleet Intelligence - Python Requirements
# For running deployment scripts locally

# Snowflake connector for Python scripts
snowflake-connector-python>=3.0.0

# Optional: For local testing of Streamlit app
# streamlit>=1.28.0
# pandas>=2.0.0
# pydeck>=0.8.0
# altair>=5.0.0
